+++ kommons-text/src/nativeMain/kotlin/com/bkahlert/kommons/text/NativeGrapheme.kt
@@ -1,7 +1,7 @@
 package com.bkahlert.kommons.text
 
 import com.github.ajalt.mordant.rendering.AnsiLevel.TRUECOLOR
-import com.github.ajalt.mordant.table.row
+import com.github.ajalt.mordant.table.horizontalLayout
 import com.github.ajalt.mordant.terminal.Terminal
 
 /** An [Iterator] that iterates [Grapheme] positions. */
@@ -56,5 +56,5 @@ private val CodePoint.isZwj get() = value == Unicode.ZWJ.code
 private val CodePoint.isSkinToneModifier get() = value in 0x1F3FB..0x1F3FF
 
 internal val terminal: Terminal by lazy { Terminal(TRUECOLOR).also { it.info.updateTerminalSize() } }
-internal val CharSequence.columns: Int get() = row { cell(this@columns) }.measure(terminal).max
+internal val CharSequence.columns: Int get() = horizontalLayout { cell(this@columns) }.measure(terminal).max
 internal val Iterable<CodePoint>.columns: Int get() = joinToString("").columns