+++ kommons/src/nativeMain/kotlin/com/bkahlert/kommons/empty.kt
@@ -1,0 +1,1 @@
+package com.bkahlert.kommons